% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{allDates}
\alias{allDates}
\title{Compute dates of leaves and internal nodes for a given tree}
\usage{
allDates(phy)
}
\arguments{
\item{phy}{Tree}
}
\value{
Dates of leaves and internal nodes
}
\description{
Compute dates of leaves and internal nodes for a given tree
}
