% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.mcmc.resBactDating}
\alias{as.mcmc.resBactDating}
\title{Convert to coda mcmc format}
\usage{
as.mcmc.resBactDating(x, burnin = 0.5)
}
\arguments{
\item{x}{Output from bactdate}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
mcmc object from coda package
}
\description{
Convert to coda mcmc format
}
