% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.treedata.resBactDating}
\alias{as.treedata.resBactDating}
\title{Convert to treedata format from ggtree package}
\usage{
as.treedata.resBactDating(x)
}
\arguments{
\item{x}{Output from bactdate}
}
\value{
treedata object from ggtree package
}
\description{
Convert to treedata format from ggtree package
}
