% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.R
\name{clusteredTest}
\alias{clusteredTest}
\title{Experimental clustered permutation test}
\usage{
clusteredTest(tree, dates, windowIncrement = 1)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{dates}{Dates of sampling}

\item{windowIncrement}{Timespan by which to increase window size}
}
\value{
Results of root-to-tip analysis after clustering
}
\description{
Experimental clustered permutation test
}
