% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coalprior}
\alias{coalprior}
\title{Coalescent prior function}
\usage{
coalprior(tab, neg)
}
\arguments{
\item{tab}{Table of nodes}

\item{neg}{Coalescent rate}
}
\value{
The log-prior in Eq (1) of Drummond et al (2002) Genetics
}
\description{
Coalescent prior function
}
