% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{coalpriorR}
\alias{coalpriorR}
\title{Coalescent prior function}
\usage{
coalpriorR(tab, neg)
}
\arguments{
\item{tab}{Table of nodes}

\item{neg}{Coalescent time unit}
}
\value{
The log-prior in Eq (1) of Drummond et al (2002) Genetics
}
\description{
Coalescent prior function
}
