% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{credate}
\alias{credate}
\title{Main function}
\usage{
credate(tree, date, initRate = 1, nbIts = 1000, useCoalPrior = T,
  updateRate = 2, initNeg = 1, updateNeg = 2, model = "poisson",
  findRoot = 0)
}
\arguments{
\item{tree}{Tree wih branches measured in unit of substitutions}

\item{date}{Sampling dates for the leaves of the tree}

\item{initRate}{Initial rate of substitutions per genome (not per site)}

\item{nbIts}{Number of MCMC iterations to perform}

\item{useCoalPrior}{Whether or not to use a coalescent prior on the tree}

\item{updateRate}{Whether or not to update the substitution rate}

\item{initNeg}{Initial rate of coalescence, equal to Ne*g}

\item{updateNeg}{Whether or not to update the neg parameter}

\item{model}{Which model to use (poisson, negbin or gamma)}

\item{findRoot}{Root finding algorithm (0=none, 1=on preset branch, 2=anywhere)}
}
\value{
Dating results
}
\description{
Main function
}
