% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{drop.tip.useRec}
\alias{drop.tip.useRec}
\title{Modified drop.tip() from 'ape' package that also updates the recombination
proportions added when loaded from CFML or Gubbins}
\usage{
drop.tip.useRec(phy, tip, trim.internal = TRUE, subtree = FALSE,
  root.edge = 0, rooted = is.rooted(phy), collapse.singles = TRUE,
  interactive = FALSE)
}
\arguments{
\item{phy}{Tree from loadCFML() or loadGubbins()}

\item{tip}{a vector of mode numeric or character specifying the tips to delete.}

\item{trim.internal}{a logical specifying whether to delete the corresponding internal branches.}

\item{subtree}{a logical specifying whether to output in the tree how many tips have been deleted and where.}

\item{collapse.singles}{a logical specifying whether to delete the internal nodes of degree 2.}

\item{node}{a node number or label.}
}
\value{
tree with rec data
}
\description{
Modified drop.tip() from 'ape' package that also updates the recombination
proportions added when loaded from CFML or Gubbins
}
