% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSample.R
\name{extractSample}
\alias{extractSample}
\title{Extract a sample of trees}
\usage{
extractSample(x, sampleSize = 500, burnin = 0.5)
}
\arguments{
\item{x}{Output from bactdate}

\item{sampleSize}{Number of trees to sample}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
list of trees
}
\description{
Extract a sample of trees
}
