% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harris2010}
\alias{harris2010}
\title{Data from Harris et al (2010)}
\usage{
data(harris2010)
}
\description{
This data is from Harris et al (2010).
}
\references{
Harris et al. (2010)
}
\keyword{datasets}
