% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{initRoot}
\alias{initRoot}
\title{Initial tree rooting based on best root-to-tip correlation}
\usage{
initRoot(phy, date)
}
\arguments{
\item{date}{Dates of sampling}

\item{tree}{Phylogenetic tree}
}
\value{
Rooted tree
}
\description{
Initial tree rooting based on best root-to-tip correlation
}
