% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood function}
\usage{
likelihood(tab, rate, n)
}
\arguments{
\item{tab}{Table of nodes}

\item{rate}{Substitution rate}

\item{n}{Number of samples}
}
\value{
log-likelihood
}
\description{
Likelihood function
}
