% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{likelihoodAcrc}
\alias{likelihoodAcrc}
\title{acrc likelihood function}
\usage{
likelihoodAcrc(tab, mu, sigma)
}
\arguments{
\item{tab}{Table of nodes}

\item{mu}{Clock rate parameter}

\item{sigma}{Std of per branch clock rate}
}
\value{
log-likelihood
}
\description{
acrc likelihood function
}
