% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{likelihoodArc}
\alias{likelihoodArc}
\title{arc likelihood function}
\usage{
likelihoodArc(tab, mu, sigma)
}
\arguments{
\item{tab}{Table of nodes}

\item{mu}{Clock rate parameter}

\item{sigma}{Std of per branch clock rate}
}
\value{
log-likelihood
}
\description{
arc likelihood function
}
