% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{likelihoodGamma}
\alias{likelihoodGamma}
\title{Gamma likelihood function}
\usage{
likelihoodGamma(tab, mu)
}
\arguments{
\item{tab}{Table of nodes}

\item{mu}{Clock rate parameter}
}
\value{
log-likelihood
}
\description{
Gamma likelihood function
}
