% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{likelihoodNegbin}
\alias{likelihoodNegbin}
\title{Negative-binomial likelihood function}
\usage{
likelihoodNegbin(tab, r, phi)
}
\arguments{
\item{tab}{Table of nodes}

\item{r}{r parameter}

\item{phi}{phi parameter}
}
\value{
log-likelihood
}
\description{
Negative-binomial likelihood function
}
