% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{likelihoodPoisson}
\alias{likelihoodPoisson}
\title{Poisson likelihood function}
\usage{
likelihoodPoisson(tab, rate)
}
\arguments{
\item{tab}{Table of nodes}

\item{rate}{Substitution rate}
}
\value{
log-likelihood
}
\description{
Poisson likelihood function
}
