% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGubbins.R
\name{loadGubbins}
\alias{loadGubbins}
\title{Load output from Gubbins}
\usage{
loadGubbins(prefix)
}
\arguments{
\item{prefix}{Prefix of the ClonalFrameML output files}
}
\value{
Tree representing the output from Gubbins
}
\description{
Load output from Gubbins
}
