% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{nodeDates}
\alias{nodeDates}
\title{Compute dates of internal nodes for a given tree}
\usage{
nodeDates(phy)
}
\arguments{
\item{phy}{Tree}
}
\value{
Dates of internal nodes
}
\description{
Compute dates of internal nodes for a given tree
}
