% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.resBactDating}
\alias{plot.resBactDating}
\title{Plotting methods}
\usage{
\method{plot}{resBactDating}(x, type = "tree", ...)
}
\arguments{
\item{x}{Output from running function credating}

\item{type}{Type of plot to do. Currently either 'tree' or 'treeCI' or 'trace' or 'treeRoot' or 'scatter'}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of BactDating results
}
\description{
Plotting methods
}
