% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.resBactDating}
\alias{print.resBactDating}
\title{Print function for resBactDating objects}
\usage{
\method{print}{resBactDating}(x, ...)
}
\arguments{
\item{x}{output from bactdate}

\item{...}{Passed on to print.phylo}
}
\value{
Print out details of BactDating results
}
\description{
Print function for resBactDating objects
}
