% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.resCreDating}
\alias{print.resCreDating}
\title{Print function for resCreDating objects}
\usage{
\method{print}{resCreDating}(x, ...)
}
\arguments{
\item{x}{output from credate}

\item{...}{Passed on to print.phylo}
}
\description{
Print function for resCreDating objects
}
