% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{relaxedLikelihood}
\alias{relaxedLikelihood}
\title{Relaxed likelihood function}
\usage{
relaxedLikelihood(tab, r, phi, n)
}
\arguments{
\item{tab}{Table of nodes}

\item{phi}{phi parameter}

\item{n}{Number of samples}

\item{rate}{r parameter}
}
\value{
log-likelihood
}
\description{
Relaxed likelihood function
}
