% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{roottotip}
\alias{roottotip}
\title{Root to tip correlation}
\usage{
roottotip(tree, date, showFig = T, predInt = F, showTree = F)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{date}{Dates of sampling}

\item{showFig}{Whether or not to show the root-to-tip regression figure}

\item{predInt}{To show 95percent confidence intervals, can be 'poisson' or 'gamma'}

\item{showTree}{Whether to show the tree or not}
}
\description{
Root to tip correlation
}
