% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simcoaltree}
\alias{simcoaltree}
\title{Simulation of coalescent dated tree}
\usage{
simcoaltree(dates = NA, alpha = 10)
}
\arguments{
\item{dates}{Vector of dates at which samples are taken}

\item{alpha}{Coalescent time unit}
}
\value{
A simulated dated tree
}
\description{
Simulation of coalescent dated tree
}
