% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simdatedtree}
\alias{simdatedtree}
\title{Simulation of dated tree using rtree and rexp branch lengths}
\usage{
simdatedtree(nsam = 20, dateroot = 2000)
}
\arguments{
\item{nsam}{Number of leaves in the tree}

\item{dateroot}{Date of the root}
}
\value{
A simulated dated tree
}
\description{
Simulation of dated tree using rtree and rexp branch lengths
}
