% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simobsphy}
\alias{simobsphy}
\title{Simulation of observed phylogeny given a dated tree}
\usage{
simobsphy(tree, rate = 10, model = "gamma")
}
\arguments{
\item{tree}{Dated tree}

\item{rate}{Substitution clock rate}

\item{model}{Which model to use (poisson or gamma)}
}
\value{
The observed phylogenetic tree
}
\description{
Simulation of observed phylogeny given a dated tree
}
