% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staph}
\alias{staph}
\title{Data on Staphylococcus aureus ST239 from Harris et al (2010)}
\usage{
data(staph)
}
\description{
This data is the Staphylococcus aureus ST239 dataset from Harris et al (2010).
}
\references{
Harris et al. (2010)
}
\keyword{datasets}
