% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.resBactDating}
\alias{summary.resBactDating}
\title{Summary function for resBactDating objects}
\usage{
\method{summary}{resBactDating}(object, ...)
}
\arguments{
\item{object}{output from bactdate}

\item{...}{Passed on to print.phylo}
}
\value{
Print out details of BactDating results
}
\description{
Summary function for resBactDating objects
}
