% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.resCreDating}
\alias{summary.resCreDating}
\title{Summary function for resCreDating objects}
\usage{
\method{summary}{resCreDating}(object, ...)
}
\arguments{
\item{object}{output from credate}

\item{...}{Passed on to print.phylo}
}
\description{
Summary function for resCreDating objects
}
