% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bactdate}
\alias{bactdate}
\title{Main function of BactDating package to date the nodes of a phylogenetic tree}
\usage{
bactdate(tree, date, initMu = NA, initAlpha = NA, initSigma = NA,
  updateMu = T, updateAlpha = T, updateSigma = T, updateRoot = T,
  nbIts = 10000, thin = ceiling(nbIts/1000), useCoalPrior = T,
  model = "mixedgamma", useRec = F, minbralen = 0.1,
  showProgress = F)
}
\arguments{
\item{tree}{Tree wih branches measured in unit of substitutions}

\item{date}{Sampling dates for the leaves of the tree}

\item{initMu}{Initial rate of substitutions per genome (not per site), or zero to use root-to-tip estimate}

\item{initAlpha}{Initial coalescent time unit}

\item{initSigma}{Initial std on per-branch substitution rate (only used in relaxed or mixed-relaxed models)}

\item{updateMu}{Whether or not to update the substitution rate}

\item{updateAlpha}{Whether or not to update the coalescent time unit}

\item{updateSigma}{Whether or not to per-branch substitution rate (only used in relaxed or mixed-relaxed models)}

\item{updateRoot}{Whether or not to use the root finding algorithm}

\item{nbIts}{Number of MCMC iterations to perform}

\item{thin}{Thining interval between recorded MCMC samples}

\item{useCoalPrior}{Whether or not to use a coalescent prior on the tree}

\item{model}{Which model to use (poisson or negbin or strictgamma or relaxedgamma or mixedgamma or arc or carc or mixedcarc)}

\item{useRec}{Whether or not to use results from previous recombination analysis}

\item{minbralen}{Minimum branch length of the phylogenetic tree (in number of substitutions)}

\item{showProgress}{Whether or not to show a progress bar}
}
\value{
Dating results
}
\description{
Main function of BactDating package to date the nodes of a phylogenetic tree
}
