% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bactdate}
\alias{bactdate}
\title{Main function}
\usage{
bactdate(tree, date, initRate = NA, initAlpha = NA, initRatevar = NA,
  updateRate = T, updateAlpha = T, updateRatevar = T, updateRoot = T,
  nbIts = 10000, thin = ceiling(nbIts/1000), useCoalPrior = T,
  model = "gamma", useRec = F, showProgress = F)
}
\arguments{
\item{tree}{Tree wih branches measured in unit of substitutions}

\item{date}{Sampling dates for the leaves of the tree}

\item{initRate}{Initial rate of substitutions per genome (not per site), or zero to use root-to-tip estimate}

\item{initAlpha}{Initial coalescent time unit}

\item{initRatevar}{Initial variance on per-branch substitution rate (only used in relaxedgamma and mixedgamma models)}

\item{updateRate}{Whether or not to update the substitution rate}

\item{updateAlpha}{Whether or not to update the coalescent time unit}

\item{updateRatevar}{Whether or not to per-branch substitution rate (only used in relaxedgamma and mixedgamma models)}

\item{updateRoot}{Whether or not to use the root finding algorithm}

\item{nbIts}{Number of MCMC iterations to perform}

\item{thin}{Thining interval between recorded MCMC samples}

\item{useCoalPrior}{Whether or not to use a coalescent prior on the tree}

\item{model}{Which model to use (poisson or gamma or relaxedgamma or mixedgamma)}

\item{useRec}{Whether or not to use results from previous recombination analysis}

\item{showProgress}{Whether or not to show a progress bar}
}
\value{
Dating results
}
\description{
Main function
}
