% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{BactDating}
\alias{BactDating}
\title{BactDating project}
\description{
This is the BactDating project.

The main functions of the package are:
\itemize{

\item roottotip
\item bactdate
}
}
\references{
Xavier Didelot, Manuscript in prepration.
}
\seealso{
https://github.com/xavierdidelot/BactDating
}
\author{
Xavier Didelot \email{xavier.didelot@gmail.com}
}
