% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.R
\name{clusteredTest}
\alias{clusteredTest}
\title{Clustered permutation test}
\usage{
clusteredTest(tree, dates)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{dates}{Dates of sampling}
}
\value{
Results of root-to-tip analysis after clustering
}
\description{
Clustered permutation test
}
