% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{credate}
\alias{credate}
\title{Main function}
\usage{
credate(tree, date, initRate = 0, nbIts = 10000,
  thin = ceiling(nbIts/1000), useCoalPrior = T, updateRate = 1,
  initNeg = 1, updateNeg = 2, initRatevar = 1, updateRatevar = 0,
  model = "gamma", updateRoot = 0, showProgress = T)
}
\arguments{
\item{tree}{Tree wih branches measured in unit of substitutions}

\item{date}{Sampling dates for the leaves of the tree}

\item{initRate}{Initial rate of substitutions per genome (not per site), or zero to use root-to-tip estimate}

\item{nbIts}{Number of MCMC iterations to perform}

\item{thin}{Thining interval between recorded MCMC samples}

\item{useCoalPrior}{Whether or not to use a coalescent prior on the tree}

\item{updateRate}{Whether or not to update the substitution rate}

\item{initNeg}{Initial rate of coalescence, equal to Ne*g}

\item{updateNeg}{Whether or not to update the neg parameter}

\item{initRatevar}{Initial variance on per-branch substituion rate (only used in relaxedgamma model)}

\item{updateRatevar}{Whether or not to per-branch substituion rate (only used in relaxedgamma model)}

\item{model}{Which model to use (poisson or negbin or gamma or relaxedgamma)}

\item{updateRoot}{Root finding algorithm (0=none, 1=on preset branch, 2=anywhere)}

\item{showProgress}{Whether or not to show a progress bar}
}
\value{
Dating results
}
\description{
Main function
}
