% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{CreDating}
\alias{CreDating}
\title{CreDating project}
\description{
This is the CreDating project.

The main functions of the package are:
\itemize{

\item creadating
\item roottotip
}
}
\references{
Xavier Didelot, Manuscript in prepration.
}
\seealso{
https://github.com/xavierdidelot/CreDating
}
\author{
Xavier Didelot \email{xavier.didelot@gmail.com}
}
