% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{credating}
\alias{credating}
\title{Main function}
\usage{
credating(tree, date, rate)
}
\arguments{
\item{tree}{Tree wih branches measured in unit of substitutions}

\item{date}{Sampling dates for the leaves of the tree}

\item{rate}{Rate per genome (not per site)}
}
\value{
Dating results
}
\description{
Main function
}
