% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{credating}
\alias{credating}
\title{Main function}
\usage{
credating(tree, date, rate = 1, nbIts = 1000, useCoalPrior = F,
  updateRate = F, neg = 1, updateNeg = F)
}
\arguments{
\item{tree}{Tree wih branches measured in unit of substitutions}

\item{date}{Sampling dates for the leaves of the tree}

\item{rate}{Rate of substitutions per genome (not per site)}

\item{nbIts}{Number of MCMC iterations to perform}

\item{useCoalPrior}{Whether or not to use a coalescent prior on the tree}

\item{updateRate}{Whether or not to update the substitution rate}

\item{neg}{Rate of coalescence, equal to Ne*g}

\item{updateNeg}{Whether or not to update the neg parameter}
}
\value{
Dating results
}
\description{
Main function
}
