% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{drop.tip.useRec}
\alias{drop.tip.useRec}
\title{Modified drop.tip() from 'ape' package that also updates the recombination
proportions added when loaded from CFML or Gubbins}
\usage{
drop.tip.useRec(
  phy,
  tip,
  trim.internal = TRUE,
  subtree = FALSE,
  root.edge = 0,
  rooted = is.rooted(phy),
  collapse.singles = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{phy}{Tree from loadCFML() or loadGubbins()}

\item{tip}{a vector of mode numeric or character specifying the tips to delete.}

\item{trim.internal}{a logical specifying whether to delete the corresponding internal branches.}

\item{subtree}{a logical specifying whether to output in the tree how many tips have been deleted and where.}

\item{root.edge}{an integer giving the number of internal branches to be used to build the new root edge. This has no effect if \code{trim.internal FALSE}.}

\item{rooted}{a logical indicating whether the tree must be treated as rooted or not. This allows to force the tree to be considered as unrooted.}

\item{collapse.singles}{a logical specifying whether to delete the internal nodes of degree 2.}

\item{interactive}{if \code{TRUE} the user is asked to select the tips or the node by clicking on the tree which must be plotted.}
}
\value{
tree with rec data
}
\description{
Modified drop.tip() from 'ape' package that also updates the recombination
proportions added when loaded from CFML or Gubbins
}
