% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{initRoot}
\alias{initRoot}
\title{Initial tree rooting based on best root-to-tip correlation}
\usage{
initRoot(phy, date, mtry = 10, useRec = F)
}
\arguments{
\item{phy}{An unrooted phylogenetic tree}

\item{date}{Dates of sampling}

\item{mtry}{Average number of rooting attempts per branch}

\item{useRec}{Whether or not to use results from previous recombination analysis}
}
\value{
Rooted tree
}
\description{
Initial tree rooting based on best root-to-tip correlation
}
