% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{leafDates}
\alias{leafDates}
\title{Compute dates of leaves for a given tree}
\usage{
leafDates(phy)
}
\arguments{
\item{phy}{Tree}
}
\value{
Dates of leaves
}
\description{
Compute dates of leaves for a given tree
}
