% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{leafDates}
\alias{leafDates}
\title{Compute dates of leaves for a given tree and date of root}
\usage{
leafDates(phy, rootdate = 0)
}
\arguments{
\item{phy}{Tree}

\item{rootdate}{Date of root}
}
\value{
Dates of leaves
}
\description{
Compute dates of leaves for a given tree and date of root
}
