% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{likelihoodNegbin}
\alias{likelihoodNegbin}
\title{Negative-binomial likelihood function}
\usage{
likelihoodNegbin(tab, mu, sigma)
}
\arguments{
\item{tab}{Table of nodes}

\item{mu}{Clock rate parameter}

\item{sigma}{Std of per branch clock rate}
}
\value{
log-likelihood
}
\description{
Negative-binomial likelihood function
}
