% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{likelihoodRelaxedgamma}
\alias{likelihoodRelaxedgamma}
\title{Relaxed gamma likelihood function}
\usage{
likelihoodRelaxedgamma(tab, rate, ratevar)
}
\arguments{
\item{tab}{Table of nodes}

\item{rate}{rate parameter}

\item{ratevar}{variance of per branch rate}
}
\value{
log-likelihood
}
\description{
Relaxed gamma likelihood function
}
