% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCFML.R
\name{loadCFML}
\alias{loadCFML}
\title{Load output from ClonalFrameML}
\usage{
loadCFML(prefix)
}
\arguments{
\item{prefix}{Prefix of the ClonalFrameML output files}
}
\description{
Load output from ClonalFrameML
}
