% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{modelcompare}
\alias{modelcompare}
\title{Perform Bayesian comparison between two models based on DIC values}
\usage{
modelcompare(res1, res2)
}
\arguments{
\item{res1}{An output from the bactdate function}

\item{res2}{Another output from the bactdate function}
}
\value{
Prints out results of model comparison
}
\description{
Perform Bayesian comparison between two models based on DIC values
}
