% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.resCreDating}
\alias{plot.resCreDating}
\title{Plotting methods}
\usage{
\method{plot}{resCreDating}(x, type = "tree", ...)
}
\arguments{
\item{x}{Output from running function credating}

\item{type}{Type of plot to do. Currently either 'tree' or 'treeCI' or 'trace'}

\item{...}{Additional parameters are passed on to plot.phylo}
}
\value{
Nothing
}
\description{
Plotting methods
}
