% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.resCreDating}
\alias{plot.resCreDating}
\title{Plotting method}
\usage{
\method{plot}{resCreDating}(res)
}
\arguments{
\item{res}{Output from running function credating}
}
\value{
Nothing
}
\description{
Plotting method
}
