% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{roottotip}
\alias{roottotip}
\title{Root to tip correlation}
\usage{
roottotip(tree, date, showTree = F)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{date}{Dates of sampling}

\item{showTree}{Whether to show the tree or not}
}
\description{
Root to tip correlation
}
