% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{roottotip}
\alias{roottotip}
\title{Root to tip correlation}
\usage{
roottotip(tree, date, permTest = 10000, showFig = T,
  showPredInt = "gamma", showTree = T)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{date}{Dates of sampling}

\item{permTest}{Number of permutations to perform to compute the p-value using a permutation test}

\item{showFig}{Whether or not to show the root-to-tip regression figure}

\item{showPredInt}{To show 95percent confidence intervals, can be 'poisson' or 'gamma'}

\item{showTree}{Whether to show the tree or not}
}
\value{
List containing estimated clock rate, date of origin and p-value
}
\description{
Root to tip correlation
}
