% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{roottotip}
\alias{roottotip}
\title{Root to tip correlation}
\usage{
roottotip(tree, date)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{date}{Dates of sampling}
}
\description{
Root to tip correlation
}
