% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{roottotip}
\alias{roottotip}
\title{Root to tip correlation}
\usage{
roottotip(
  tree,
  date,
  rate = NA,
  permTest = 10000,
  showFig = T,
  colored = T,
  showPredInt = "gamma",
  showText = T,
  showTree = T
)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{date}{Dates of sampling}

\item{rate}{Evolutionary rate, estimated unless provided}

\item{permTest}{Number of permutations to perform to compute the p-value using a permutation test}

\item{showFig}{Whether or not to show the root-to-tip regression figure}

\item{colored}{Whether or not to use colors illustrating dates}

\item{showPredInt}{To show 95percent confidence intervals of a strict clock model, can be 'poisson' or 'gamma'}

\item{showText}{Whether to show the title and axis labels}

\item{showTree}{Whether to show the tree or not}
}
\value{
List containing estimated clock rate, date of origin and p-value
}
\description{
Root to tip correlation
}
