% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roototip.R
\name{roottotip}
\alias{roottotip}
\title{Root to tip correlation}
\usage{
roottotip(tree, date, permTest = 10000, showFig = T,
  showPredInt = "gamma", showText = T, showTree = T, show.tip.label = F)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{date}{Dates of sampling}

\item{permTest}{Number of permutations to perform to compute the p-value using a permutation test}

\item{showFig}{Whether or not to show the root-to-tip regression figure}

\item{showPredInt}{To show 95percent confidence intervals, can be 'poisson' or 'gamma'}

\item{showText}{Whether to show the title and axis labels}

\item{showTree}{Whether to show the tree or not}

\item{show.tip.label}{Whether or not to show tip labels on the tree}
}
\value{
List containing estimated clock rate, date of origin and p-value
}
\description{
Root to tip correlation
}
