% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simcoaltree}
\alias{simcoaltree}
\title{Simulation of coalescent dated tree}
\usage{
simcoaltree(dates, neg)
}
\arguments{
\item{neg}{Reciprocal of coalescent rate}

\item{times}{dates at which samples are taken}
}
\value{
A simulated dated tree
}
\description{
Simulation of coalescent dated tree
}
