% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simobsphy}
\alias{simobsphy}
\title{Simulation of observed phylogeny given a dated tree}
\usage{
simobsphy(tree, mu = 10, sigma = 0, model = "gamma")
}
\arguments{
\item{tree}{Dated tree}

\item{mu}{Substitution clock rate}

\item{sigma}{Per-branch std on the clock rate (used only by relaxed gamma model)}

\item{model}{Which model to use (poisson or gamma or relaxedgamma)}
}
\value{
An observed phylogenetic tree
}
\description{
Simulation of observed phylogeny given a dated tree
}
